Platform-specific fuse support for Linux, MacOS
