+++ weed/command/fuse_notsupported.go
@@ -1,0 +1,15 @@
+//go:build !linux && !darwin
+// +build !linux,!darwin
+
+package command
+
+import (
+	"fmt"
+	"runtime"
+)
+
+func runFuse(cmd *Command, args []string) bool {
+	fmt.Printf("Fuse is not supported on %s %s\n", runtime.GOOS, runtime.GOARCH)
+
+	return true
+}
